requirements.txt
mlx
safetensors
huggingface-hub
regex
numpy
tqdm
Pillow
streamlit
diffusers
omegaconf